% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occbin.R
\name{.forward_simulate}
\alias{.forward_simulate}
\title{Forward simulate a linear state-space system}
\usage{
.forward_simulate(H, G, M, x0, shock_path, horizon)
}
\description{
Forward simulate a linear state-space system
}
\keyword{internal}
