% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{forecast}
\alias{forecast}
\title{Forecast from a DSGE Model}
\usage{
forecast(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A forecast object.
}
\description{
Generic function for forecasting from estimated DSGE models.
}
\examples{
\donttest{
m <- dsge_model(
  obs(y ~ z),
  state(z ~ rho * z),
  start = list(rho = 0.5)
)
set.seed(42)
z <- numeric(150)
for (i in 2:150) z[i] <- 0.8 * z[i - 1] + rnorm(1)
fit <- estimate(m, data = data.frame(y = z))
fc <- forecast(fit, horizon = 8)
plot(fc)
}

}
\seealso{
\code{\link[=forecast.dsge_fit]{forecast.dsge_fit()}}
}
