% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-specification.R
\name{obs}
\alias{obs}
\title{Define an Observed Control Variable Equation}
\usage{
obs(formula)
}
\arguments{
\item{formula}{A formula of the form \code{variable ~ expression}.}
}
\value{
A list with class \code{"dsge_equation"} containing the parsed equation
and its type.
}
\description{
Wraps a formula to mark it as an equation for an observed control variable
in a linear DSGE model.
}
\examples{
obs(y ~ z)

}
\seealso{
\code{\link[=unobs]{unobs()}}, \code{\link[=state]{state()}}, \code{\link[=dsge_model]{dsge_model()}}
}
