% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoother.R
\name{plot.dsge_decomposition}
\alias{plot.dsge_decomposition}
\title{Plot Historical Shock Decomposition}
\usage{
\method{plot}{dsge_decomposition}(x, which = NULL, ...)
}
\arguments{
\item{x}{A \code{dsge_decomposition} object.}

\item{which}{Which observable(s) to plot. Integer or character. Default
is all.}

\item{...}{Additional arguments (currently unused).}
}
\value{
No return value, called for the side effect of producing
stacked bar charts of the historical shock decomposition on the
active graphics device.
}
\description{
Creates a stacked bar chart showing the contribution of each structural
shock to the observed variables over time.
}
\examples{
\donttest{
m <- dsge_model(
  obs(y ~ z),
  state(z ~ rho * z),
  start = list(rho = 0.5)
)
set.seed(42)
z <- numeric(150)
for (i in 2:150) z[i] <- 0.8 * z[i - 1] + rnorm(1)
fit <- estimate(m, data = data.frame(y = z))
plot(shock_decomposition(fit))
}

}
