% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-tools.R
\name{fitted.dsge_fit}
\alias{fitted.dsge_fit}
\title{Fitted values from a DSGE model}
\usage{
\method{fitted}{dsge_fit}(object, ...)
}
\arguments{
\item{object}{A \code{"dsge_fit"} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A matrix of fitted values with columns named by observed variables.
}
\description{
Returns filtered fitted values of observed variables (using all
information up to and including time t).
}
