% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot.dsge_forecast}
\alias{plot.dsge_forecast}
\title{Plot DSGE Forecasts}
\usage{
\method{plot}{dsge_forecast}(x, ...)
}
\arguments{
\item{x}{A \code{dsge_forecast} object from \code{\link[=forecast.dsge_fit]{forecast.dsge_fit()}}.}

\item{...}{Additional arguments passed to base plotting functions.}
}
\value{
No return value, called for the side effect of producing
forecast path plots on the active graphics device.
}
\description{
Plots forecast paths for observed variables.
}
