% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/second-order.R
\name{plot.dsge_irf_2nd}
\alias{plot.dsge_irf_2nd}
\title{Plot Second-Order Impulse Responses}
\usage{
\method{plot}{dsge_irf_2nd}(x, variables = NULL, ...)
}
\arguments{
\item{x}{A \code{dsge_irf_2nd} object from \code{\link[=irf_2nd_order]{irf_2nd_order()}}.}

\item{variables}{Optional character vector of variables to plot.
Default: all variables.}

\item{...}{Additional arguments passed to base plotting functions.}
}
\value{
No return value, called for the side effect of producing the
plots on the active graphics device.
}
\description{
Plots the responses from \code{\link[=irf_2nd_order]{irf_2nd_order()}}, one panel per variable.
}
\examples{
rbc <- read_dynare(system.file("examples", "rbc.mod", package = "dsge"))
sol2 <- solve_dsge(rbc, order = 2)
plot(irf_2nd_order(sol2, shock = "e", size = 0.05), variables = c("c", "k"))

}
